Revolutio/requirements/base_others4.txt
azure-mgmt-storage
azure-storage-blob
boto3==1.34.41
Collectfast==2.2.0
connectorx==0.3.2
crispy-bootstrap4==2023.1
cssmin==0.2.0
dask[complete]==2024.2.0
django-auth-ldap==4.6.0
django-cors-headers==4.3.1
django-prometheus==2.3.1
django-storages==1.14.2
django_multitenant==4.1.1
djangorestframework-simplejwt
docker==7.0.0
jsmin==3.0.1
ldap3
msoffcrypto-tool==5.3.1
numpy-financial==1.0.0
oracledb==2.0.1
paramiko==3.4.0
pybaht==0.2.0
pycryptodome==3.20.0
pyotp==2.9.0
PyPDF2==3.0.1
python-ldap==3.4.4
pyxirr
qrcode==7.4.2
reportlab==4.1.0
rq-scheduler==0.13.1
weasyprint==61.0
xgboost
xlrd==2.0.1
xlsxwriter==3.1.9
xmltodict==0.13.0
yearfrac==0.4.8
zope.interface==6.1
